Mark unused vars as comment
git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@104 e54a6e83-5905-42d5-8d5c-058d10e6a962

+++ lib/connection/basicclient.cpp
@@ -112,7 +112,7 @@ bool CBasicClient::send_string(const char* data)
 	}
 	else
 	{
-		send_length = length;
+		send_length = static_cast<uint8_t>(length);
 	}
 	return (send_data((char *)&send_length, sizeof(send_length)) &&
 		send_data(data, send_length));
